complete student endpoints tests
